backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
lxml==4.9.3
fake-useragent==1.4.0
unidecode==1.3.7
phonenumbers==8.13.25
webdriver-manager==4.0.1
chromedriver-autoinstaller==0.6.2
stripe==6.6.0 